- Remove unsigned-signed comparison warnings.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@5345 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/jrd.h
@@ -739,7 +739,7 @@ public:
 		{
 			if (!base)
 				base = FB_NEW(p) vec(p, len);
-			else if (len > base->count())
+			else if (len > (int) base->count())
 				base->resize(len);
 			return base;
 		}
@@ -761,7 +761,7 @@ public:
 		{
 			if (!base)
 				base = FB_NEW(p) vcl(p, len);
-			else if (len > base->count())
+			else if (len > (int) base->count())
 				base->resize(len);
 			return base;
 		}
